requirements.txt
opencv-python==4.5.4.60
Pillow==8.4.0
scikit-image==0.19.0
scipy==1.7.3
torch==1.7.1
tqdm==4.62.3
wandb==0.12.10